Desktop/Turk_uzb/requirements.txt
pyTelegramBotAPI
deep-translator
gTTS
Flask
python-dotenv
